Fix slidingView position
The original implementation sets the slidingView position based on the
firstTrigger value.
However, trigger values are configurable, and calculating the
slidingView position based on the firstTrigger value does not seem to a
good idea especially when the customized firstTrigger value is very
small.
Therefore, the new implementation sets the slidingView position based on
the fixed value, 0.25, which is the default firstTrigger value.